Aula 3 - Integrando com SOAP dos Correios

+++ controllers/pagamento.js
@@ -1,3 +1,7 @@
+const PAGAMENTO_CRIADO = "CRIADO";
+const PAGAMENTO_CONFIRMADO = "CONFIRMADO";
+const PAGAMENTO_CANCELADO = "CANCELADO";
+
 module.exports= (app)=>{
     app.get('/pagamentos', (req, res)=>{
         res.send('OK');
@@ -6,10 +10,10 @@ module.exports= (app)=>{
     app.delete('/pagamentos/pagamento/:id', (req, res)=>{
         
         let pagamento = {};
-        let id = req.params.id;
+        let id = req.para.id;
 
         pagamento.id =id;
-        pagamento.status = 'CANCELADO'
+        pagamento.status = PAGAMENTO_CANCELADO;
 
         let connection = app.persistencia.connectionFactory();
         let pagamentoDao = new app.persistencia.PagamentoDao(connection);
@@ -31,7 +35,7 @@ module.exports= (app)=>{
         let id = req.params.id;
 
         pagamento.id =id;
-        pagamento.status = 'CONFIRMADO'
+        pagamento.status = PAGAMENTO_CONFIRMADO;
 
         let connection = app.persistencia.connectionFactory();
         let pagamentoDao = new app.persistencia.PagamentoDao(connection);
@@ -49,8 +53,8 @@ module.exports= (app)=>{
 
     app.post('/pagamentos/pagamento', (req, res)=>{
         
-        req.assert("forma_de_pagamento", "Forma de pagamento é obrigatório").notEmpty();
-        req.assert("valor", "Valor obrigatório e deve ser um decimal").notEmpty().isFloat();
+        req.assert("pagamento.forma_de_pagamento", "Forma de pagamento é obrigatório").notEmpty();
+        req.assert("pagamento.valor", "Valor obrigatório e deve ser um decimal").notEmpty().isFloat();
 
         let erros = req.validationErrors();
 
@@ -60,10 +64,10 @@ module.exports= (app)=>{
             return;
         }
         
-        let pagamento = req.body;
+        let pagamento = req.body["pagamento"];
         console.log('Processando uma requisição de um novo pagamento');
 
-        pagamento.status = 'CRIADO';
+        pagamento.status = PAGAMENTO_CRIADO;
         pagamento.data = new Date;
 
         let connection = app.persistencia.connectionFactory();
@@ -78,25 +82,52 @@ module.exports= (app)=>{
             }else{
                 pagamento.id = resultado.insertId;
                 console.log('Pagamento criado');
-                res.location('/pagamentos/pagamento/' + pagamento.id);
-                
-                let response = {
-                    dados_do_pagamento: pagamento,
-                    links: [
-                        {
-                            href:'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
-                            rel: 'confirmar',
-                            method: 'PUT'
-                        },
-                        {
-                            href:'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
-                            rel: 'cancelar',
-                            method: 'DELETE'
+
+                if(pagamento.forma_de_pagamento == 'cartao'){
+                    let cartao = req.body['cartao'];
+                    console.log(cartao);
+
+                    let clienteCartoes = new app.servicos.clienteCartoes();
+
+                    clienteCartoes.autoriza(cartao, (err, request, response, retorno)=>{
+                        
+                        if (err) {
+                            console.log(err);
+                            
+                            res.status(400).json(err);
+                            return;
+                            
                         }
-                    ]
-                }
+
+                        console.log(retorno);
+                        res.status(201).json(retorno);
+                        return;
+                    });
+
+
+                } else {
+
+                    res.location('/pagamentos/pagamento/' + pagamento.id);
                 
-                res.status(201).json(response)
+                    let response = {
+                        dados_do_pagamento: pagamento,
+                        links: [
+                            {
+                                href:'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
+                                rel: 'confirmar',
+                                method: 'PUT'
+                            },
+                            {
+                                href:'http://localhost:3000/pagamentos/pagamento/' + pagamento.id,
+                                rel: 'cancelar',
+                                method: 'DELETE'
+                            }
+                        ]
+                    }
+                    
+                    res.status(201).json(response)
+                }
+
             }            
             
         });
